Factor out sendable_email.py.
